+++ src/felupe/mechanics/__init__.py
@@ -12,7 +12,7 @@ from ._solidbody_gravity import SolidBodyGravity
 from ._solidbody_incompressible import SolidBodyNearlyIncompressible
 from ._solidbody_pressure import SolidBodyPressure
 from ._step import Step
-from ._truss import Truss
+from ._truss import TrussBody
 
 __all__ = [
     "Assemble",
@@ -33,5 +33,5 @@ __all__ = [
     "Step",
     "MultiPointConstraint",
     "MultiPointContact",
-    "Truss",
+    "TrussBody",
 ]